move protobuf serializer to its own package so only users need to take a protobuf dep
