+++ fsm/protobufserializer/serializer_test.go
@@ -1,0 +1,61 @@
+package protobufserializer
+
+import (
+	"testing"
+
+	"github.com/golang/protobuf/proto"
+	. "github.com/sclasen/swfsm/log"
+)
+
+func TestProtobufSerialization(t *testing.T) {
+	ser := &ProtobufStateSerializer{}
+	key := "FOO"
+	val := "BAR"
+	init := &ConfigVar{Key: &key, Str: &val}
+	serialized, err := ser.Serialize(init)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	Log.Println(serialized)
+
+	deserialized := new(ConfigVar)
+	err = ser.Deserialize(serialized, deserialized)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if init.GetKey() != deserialized.GetKey() {
+		t.Fatal(init, deserialized)
+	}
+
+	if init.GetStr() != deserialized.GetStr() {
+		t.Fatal(init, deserialized)
+	}
+}
+
+//This is c&p from som generated code
+
+type ConfigVar struct {
+	Key             *string `protobuf:"bytes,1,req,name=key" json:"key,omitempty"`
+	Str             *string `protobuf:"bytes,2,opt,name=str" json:"str,omitempty"`
+	XXXunrecognized []byte  `json:"-"`
+}
+
+func (m *ConfigVar) Reset()         { *m = ConfigVar{} }
+func (m *ConfigVar) String() string { return proto.CompactTextString(m) }
+func (*ConfigVar) ProtoMessage()    {}
+
+func (m *ConfigVar) GetKey() string {
+	if m != nil && m.Key != nil {
+		return *m.Key
+	}
+	return ""
+}
+
+func (m *ConfigVar) GetStr() string {
+	if m != nil && m.Str != nil {
+		return *m.Str
+	}
+	return ""
+}
